# openapi.yaml
openapi: 3.0.3
...
info:
  title: Ryden REST API
  version: "1.0"
  description: >
    API REST (Jersey/JAX-RS) de Ryden — plataforma P2P de alquiler de autos.
    Reemplaza la capa Spring MVC + JSP. Consumida por una SPA servida como
    estáticos desde Tomcat (sin SSR). Esta especificación es el contrato
    público (DTOs), independiente del modelo de dominio.
servers:
  - url: /webapp
    description: >
      Context path en Tomcat local (WAR desplegado como webapp.war).
      En Jetty dev (`mvn jetty:run`) el context path es `/` — ajustar el prefijo al probar.
  - url: /
    description: Raíz de la API en Jetty dev o despliegue sin context path
...
security:
  - bearerAuth: []
paths:
...
  /users/{id}/documents/{documentType}:
    parameters:
      - $ref: "#/components/parameters/idPath"
      - name: documentType
        in: path
        required: true
        schema: { type: string, enum: [license, identity] }
    get:
      tags: [users, files]
      summary: Descargar documento de perfil (self/admin)
      responses:
        "200": { $ref: "#/components/responses/BinaryFile" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [users, files]
      summary: Subir/reemplazar un documento (self)
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: { type: string, format: binary }
      responses:
        "204": { description: Subido. }
        "400": { $ref: "#/components/responses/ValidationError" }
        "403": { $ref: "#/components/responses/Forbidden" }
    delete:
      tags: [users, files]
      summary: Borrar un documento (self)
      responses:
        "204": { description: Borrado. }
        "403": { $ref: "#/components/responses/Forbidden" }

  # --- Favoritos (entidad débil FavCar; colección privada del usuario) ---
...
  /reservations/{id}:
    parameters:
      - $ref: "#/components/parameters/idPath"
    get:
      tags: [reservations]
      summary: Detalle de una reserva (rider/owner/admin)
      responses:
        "200":
          description: Reserva.
          content:
            application/vnd.paw.reservation.v1+json:
              schema: { $ref: "#/components/schemas/ReservationDto" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [reservations]
      summary: Transiciones de estado de la reserva (accept/cancel/return/finish + editar fechas)
      description: >
        Absorbe TODAS las acciones del MVC sobre la reserva:
        - Editar fechas de una PENDING impaga (rider) → { "startDate", "endDate" }.
        - Cancelar (rider u owner) → { "status": "cancelled_by_rider" |
          "cancelled_by_owner" }. Reemplaza /cancel. Genera obligación de
          refund si correspondía.
        - Marcar auto devuelto (owner) → { "carReturned": true }. Reemplaza
          /car-returned. Puede transicionar a FINISHED.
        - (La transición PENDING→ACCEPTED ocurre al subir el comprobante de
          pago — ver PUT .../payment-receipt — no por PATCH manual.)
        El resource valida la transición permitida según estado y rol
        (403/409 según corresponda).
      requestBody:
        required: true
        content:
          application/vnd.paw.reservation.v1+json:
            schema: { $ref: "#/components/schemas/ReservationPatchDto" }
      responses:
        "200":
          description: Reserva actualizada.
          content:
            application/vnd.paw.reservation.v1+json:
              schema: { $ref: "#/components/schemas/ReservationDto" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Transición no permitida desde el estado actual.
...
  /reservations/{id}/reviews:
    parameters:
      - $ref: "#/components/parameters/idPath"
    get:
      tags: [reviews, reservations]
      summary: Reseñas de la reserva (0, 1 o 2)
      responses:
        "200":
          description: Reseñas de la reserva.
          content:
            application/vnd.paw.review.v1+json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ReviewDto" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [reviews, reservations]
      summary: Dejar una reseña (rider→owner u owner→rider)
      description: >
        Reemplaza /owner-review-rider y /rider-review-owner. El lado (made_by_rider)
        se deriva del rol del llamante en la reserva. Solo si la reserva está
        FINISHED y no existe ya una reseña de ese lado (UNIQUE(reservation_id, made_by_rider)).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                rating: { type: integer, minimum: 1, maximum: 5 }
                comment: { type: string }
                image: { type: string, format: binary }
            encoding:
              image: { contentType: image/* }
      responses:
        "201":
          description: Reseña creada.
          headers:
            Location: { $ref: "#/components/headers/Location" }
          content:
            application/vnd.paw.review.v1+json:
              schema: { $ref: "#/components/schemas/ReviewDto" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: Ya existe reseña de ese lado, o la reserva no está FINISHED.
  /reservations/{id}/reviews/{reviewId}:
    parameters:
      - $ref: "#/components/parameters/idPath"
      - name: reviewId
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      tags: [reviews, reservations]
      summary: Reseña individual por su id propio (URN única)
      description: >
        Recurso canónico de una Review — es el mismo destino de links.self tanto
        en ReviewDto (/cars/{id}/reviews, /reservations/{id}/reviews) como en
        UserReviewDto (/users/{id}/reviews).
      responses:
        "200":
          description: La reseña.
          content:
            application/vnd.paw.review.v1+json:
              schema: { $ref: "#/components/schemas/ReviewDto" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  #############################################################################
  # RESERVATION MESSAGES (chat) — entidad débil
  #############################################################################
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: >
        Authorization: Basic base64(email:secret) a CUALQUIER endpoint.
        El secret depende del contexto:
        - Cuenta verificada: password (BCrypt).
        - Verificación de email (cuenta sin verificar): OTP de 6 dígitos enviado
          por mail tras POST /users. Se consume al autenticar; devuelve JWT.
        - Reset de password olvidado: OTP enviado tras POST /credentials. Otorga
          authority ROLE_PASSWORD_RESET_OTP; se consume al completar PATCH
          /users/{id} con password + passwordConfirm.
        Tras auth válida el server responde con X-Access-Token, X-Refresh-Token
        y Link rel="authenticated-user". No hay /login ni /verify.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Requests normales: Authorization: Bearer <access_token>. Cuando vence
        (401), se manda el Refresh Token (también como Bearer) a cualquier
        endpoint y el server devuelve un par nuevo en headers. No hay /refresh:
        el JWT es autocontenido y el server distingue access de refresh.
  parameters:
    idPath:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
  headers:
    Location:
      description: URN canónica del recurso recién creado.
      schema: { type: string, format: uri }
...
  responses:
...
    Forbidden:
      description: 403 — Autenticado pero sin permiso.
      content:
        application/vnd.paw.error.v1+json:
          schema: { $ref: "#/components/schemas/ErrorDto" }
    NotFound:
      description: 404 — El recurso no existe.
      content:
        application/vnd.paw.error.v1+json:
          schema: { $ref: "#/components/schemas/ErrorDto" }
...
    ValidationError:
      description: 400 — Request mal formado o validación fallida.
      content:
        application/vnd.paw.error.v1+json:
          schema: { $ref: "#/components/schemas/ValidationErrorDto" }
...
    BinaryFile:
      description: Bytes de un archivo (pdf/imagen/video).
      content:
        application/octet-stream:
          schema: { type: string, format: binary }
  schemas:
    Links:
      type: object
      description: Bloque de hipervínculos generado on-the-fly. Siempre incluye self. NUNCA se persiste.
      additionalProperties: { type: string, format: uri }
      properties:
        self: { type: string, format: uri }
      required: [self]
...
    ReservationStatus:
      type: string
      enum:
        - pending
        - accepted
        - started
        - cancelled
        - cancelled_by_rider
        - cancelled_by_owner
        - cancelled_due_to_missing_payment_proof
        - finished
    ReservationDto:
      type: object
      properties:
        startDate: { type: string, format: date-time, description: UTC. }
        endDate: { type: string, format: date-time, description: UTC. }
        status: { $ref: "#/components/schemas/ReservationStatus" }
        totalPrice: { type: number }
        carReturned: { type: boolean }
        paymentProofDeadlineAt: { type: string, format: date-time, nullable: true }
        refundProofDeadlineAt: { type: string, format: date-time, nullable: true }
        paymentRefundRequired: { type: boolean }
        hasPaymentReceipt: { type: boolean }
        hasRefundReceipt: { type: boolean }
        ownerCbu:
          type: string
          nullable: true
          description: CBU del dueño del auto, para que el rider transfiera el pago.
        pickupStreet: { type: string, nullable: true, description: Calle de entrega/devolución. }
        pickupNumber: { type: string, nullable: true }
        pickupNeighborhood: { type: string, nullable: true }
        checkInTime: { type: string, nullable: true, example: "10:00" }
        checkOutTime: { type: string, nullable: true, example: "18:00" }
        createdAt: { type: string, format: date-time }
        links:
          allOf:
            - $ref: "#/components/schemas/Links"
          description: self, car, rider, owner, messages, reviews, payment-receipt, refund-receipt.
...
    ReservationPatchDto:
      type: object
      properties:
        status: { $ref: "#/components/schemas/ReservationStatus" }
        carReturned: { type: boolean }
        startDate: { type: string, format: date-time, description: Solo PENDING impaga (rider). }
        endDate: { type: string, format: date-time }
...
    ReviewDto:
      type: object
      description: >
        Review es una entidad fuerte (id propio); (reservation_id, made_by_rider)
        queda como UNIQUE para la regla de "una reseña por lado", ya no como PK.
        Esto le da a cada reseña una URN única (links.self), aunque se la liste
        desde /cars/{id}/reviews, /users/{id}/reviews o /reservations/{id}/reviews.
      properties:
        rating: { type: integer, minimum: 1, maximum: 5, nullable: true }
        comment: { type: string, nullable: true }
        madeByRider: { type: boolean, description: true = reseña del rider hacia el owner. }
        createdAt: { type: string, format: date-time }
        links:
          allOf:
            - $ref: "#/components/schemas/Links"
          description: >
            self (único, /reservations/{id}/reviews/{reviewId}), reservation, car,
            author, image (si hay).
...
    ErrorDto:
      type: object
      properties:
        status: { type: integer }
        code: { type: string }
        message: { type: string }
    ValidationErrorDto:
      type: object
      properties:
        status: { type: integer, example: 400 }
        message: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }